+++ manim_slides/convert.py
@@ -421,6 +421,8 @@ class RevealJS(Converter):
             for presentation_config in self.presentation_configs:
                 presentation_config.copy_to(full_assets_dir)
 
+        dest.parent.mkdir(parents=True, exist_ok=True)
+
         with open(dest, "w") as f:
             sections = "".join(self.get_sections_iter(assets_dir))
 
@@ -480,6 +482,8 @@ class PDF(Converter):
                 images.append(
                     read_image_from_video_file(slide_config.file, self.frame_index)
                 )
+
+        dest.parent.mkdir(parents=True, exist_ok=True)
 
         images[0].save(
             dest,
@@ -571,6 +575,7 @@ class PowerPoint(Converter):
                 if self.auto_play_media:
                     auto_play_media(movie, loop=slide_config.is_loop())
 
+        dest.parent.mkdir(parents=True, exist_ok=True)
         prs.save(dest)